Moved DayFragment to its own file, sorted in EventDao, put text width limit on list items
