+++ app/src/main/java/org/hackillinois/android/database/dao/EventDao.kt
@@ -15,7 +15,7 @@ interface EventDao {
     @Query("SELECT * FROM events WHERE startTime <= :time AND endTime >= :time")
     fun getAllEventsHappeningAtTime(time: Long): LiveData<List<Event>>
 
-    @Query("SELECT * FROM events WHERE startTime >= :startTime AND startTime < :endTime")
+    @Query("SELECT * FROM events WHERE startTime >= :startTime AND startTime < :endTime ORDER BY startTime")
     fun getEventsHappeningBetweenTimes(startTime: Long, endTime: Long): LiveData<List<Event>>
 
     @Query("SELECT * FROM events WHERE last_refreshed > :lastRefreshMax AND name LIKE :name LIMIT 1")
